change amount_per_round to decimal

+++ programs/exchange/src/lib.rs
@@ -1888,7 +1888,7 @@ pub struct StakingRound {
 pub struct Staking {
     pub fund_account: Pubkey,         //32 Source account of SNY tokens
     pub round_length: u32,            //4 Length of round in slots
-    pub amount_per_round: u64,        //8 Amount of SNY distributed per round
+    pub amount_per_round: Decimal,    //8 Amount of SNY distributed per round
     pub finished_round: StakingRound, //24
     pub current_round: StakingRound,  //24
     pub next_round: StakingRound,     //24
